DjangoVrboBackend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-channels==0.7.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
Markdown==3.7
oauthlib==3.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
